Added a column component
