+++ public/css/kanban.css
@@ -5,3 +5,10 @@
 .card {
     margin: 10px;
 }
+
+.kanban-column {
+    min-height: 100px;
+    padding: 5px;
+    border: 3px solid;
+    border-radius: 5px;
+}
